requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.1
openai>=1.13.0
requests>=2.31.0
typing-extensions>=4.8.0 